fix(shared-data,app): Correctly ignore certain offsets in a labware stack (#19466)
Closes RQA-4637.

We're trying to unify all of our frontend labware positioning math to a
single place, and we're trying to get that single place to behave just
like the backend, where we're pursuing similar unification.

In my crusade to do that, I made a bad assumption about where labware
ends up when it's stacked. Labware definitions define offsets relative
to each other, and I thought that you could sum up all of those offsets
to compute the final value. That's probably how it *ought* to work, but
it isn't, actually. In fact, the intermediate labwares basically only
contribute to the final z offset, not the x or y offsets. Tip racks and
adapters have been relying on this, so we can't change it.

This patches the math to emulate that oddity. It's kind of kludgy,
sorry.

This also cherry-picks #19431.