feat(shared-data): restore liquid class `water` v2 definitions (#19436)
# Overview

Now that we have versioning support in `get_liquid_class()` (PR #19408),
this PR takes the updated liquid class values for `water` (PR #19141)
and makes them available as version 2 of the `water` class. AUTH-2246

This PR also makes the preliminary values for the low-volume 96
`flex_96channel_200` pipette available in `water` v2, so that they can
be used by hardware testing. If we decide that we don't want to release
`flex_96channel_200` in the public RS 8.6.1 release, we'll delete the
`flex_96channel_200` values again.

And this fixes the `test_gravimetric_protocol.py` test to stop failing.
It's been failing since PR #19306 when we deleted the
`flex_96channel_200` liquid class values from RS 8.6.0.

## Test Plan and Hands on Testing

Ran the updated `test_gravimetric_protocol.py` and confirmed that it
passes.

Confirmed that `test_protocol_core.py` still passes.

## Risk assessment

Low. But remember to remove the `flex_96channel_200` values from `water`
v2 if we don't want to release them.