feat(api): Publicize `opentrons.protocol_api.OffDeckType` (#19484)
## Overview

This exposes `OffDeckType`, which is the type of the special `OFF_DECK`
constant, as a public interface of the Python Protocol API. This allows
protocol authors to use it in type annotations. Closes #19473 /
AUTH-2261.

## Changelog

* Allow importing `OffDeckType` from `opentrons.protocol_api`.
* Include `OffDeckType` in the reference documentation.

I'm not gating anything behind `apiLevel` because `apiLevel` in general
can't really control things like this, having to do with
`opentrons.protocol_api` import structure. For comparison, when the
`OFF_DECK` constant was introduced, we couldn't gate its import behind
`apiLevel`—we could only gate its usage in versioned `ProtocolContext`
methods.