test(protocol-designer): Execute Cypress tests eagerly (#19505)
This tries to make it easier to debug failing Cypress tests.

Currently, the tests follow a builder pattern, like this:

```javascript
// Just add the steps to a list, without actually running them.
steps.add(Foo())
steps.add(Bar())

// Actually run all the steps.
steps.execute()
```

The downside of that is that, when a step fails, the stack trace
opaquely points to the for-loop inside `execute()`.

At a glance, it didn't seem like the builder pattern was doing anything
for us, so this PR just deletes it in favor of executing the steps
eagerly. This way, when a step fails, the stack trace points to the
actual failing step in the `.cy.ts` file.