feat(protocol-designer): Update StepContainer for concurrent module actions (#19477)
## Overview

This implements design changes to Protocol Designer's `StepContainer`
component, to support concurrent module actions. Closes EXEC-1688—see
there for a link to the designs.

## Changelog

* `StepContainer` can now optionally have a second line of text. (The
"subtext.")
* `StepContainer`'s height is no longer hard-coded. Instead, it follows
the height of its children, matching how it works in Figma. This ends up
a little bit weird because the icon for the overflow menu needs to
contribute to the height even when it isn't present.