docs: add Stacker summary to Flex manual (#19515)
# Overview

This adds a Stacker section to the Flex manual, Modules chapter. It
provides a brief summary of the Stacker and links to the instruction
manual.

RTC-765

# Overview

A short piece that summarizes the purpose of the Stacker module,
includes links to more thorough documentation.

## Test Plan and Hands on Testing

Checked links to make sure they work.

## Changelog

- Adds a new file `stacker.md` to `modules`.
- Adds new text and updates the table for `modules > index.md`.

## Review requests

I'm stumped on if or how (even where) to add some text that all labware
in the Stacker must be uniform, meaning although it can hold up to 6 tip
racks, for example, they all have to be the same capacity and have or do
not have lids. You can't mix 'em up. I end up with really tortured prose
here so need some help/advice.

## Risk assessment

Low, docs only.

---------

Co-authored-by: Ed Cormany <edward.cormany@opentrons.com>