feat(api): improve gripper error checking
When we move labware during the gripper, we use the position of the
labware jaw while labware is nominally gripped to determine whether we
successfully grabbed the labware or not. We compare the position to some
fact of the labware's geometry.

Because the geometry doesn't actually contain the width at the point
that it's gripped (it only has the bounding box, which includes e.g. a
skirt that the gripper doesn't interact with) and because the gripper's
jaw width position tracking can have problems, we also allow
uncertainty.

There were a couple problems that are now fixed, such as
- We were doing an exact position comparison for the
gripper-fully-closed case; this is not really safe because the encoder
can be a little off, so add a less-than check to make sure we detect it
- The big one: we determined our uncertainty values by the difference
between the well bounding box and the labware bounding box. For labware
that doesn't have wells, that means the uncertainty is effectively
infinite. Instead, use a fixed distance.
- Also, refactor some stuff to support it.

Closes EXEC-1783