feat(api): cancel protocols for asynch module errors (#19503)
Sometimes, our modules notice that an error has occurred while they're
doing ongoing and generally less-monitored work like maintaining a
static temperature. This work is usually not tied to a specific command
or task (and still won't be after we have concurrent module actions,
because the concurrent module action "preheat" is considered to end once
the module has reached its temperature, but the module is still
controlling at that temperature thereafter). The errors are usually
noticed by the modules' status pollers, which are always running in the
background while the modules are connected to make sure that callers
like the app have up-to-date status data.

Today, when there's an error noticed only by the status poller,
essentially nothing happens; the error doesn't go anywhere. Usually, the
status poller just fails and the module is dead, but nothing will notice
until and unless it sends another module status request that might get
the module hardware to respond with the ongoing error (typically, if a
module has a subsystem error it will respond with an error code to
commands that affect that subsystem). If there is no such command, the
module failure goes unnoticed.

To fix this, take the hardware callback mechanism that we use today for
estop and for door status and extend it to asynchronous module errors.

There are two beefy commits that each implement half the work: extending
the hardware error callback up through the server and protocol engine
happens in 3c252e72bbd10cf1ec0f270e3beecfe6cc3fba3c , and extending the
hardware error callback down to the module pollers happens in
a63d3892001e0c4f90cd7dc614bfc5f465180eda .

## extending up through the engine
3c252e72bbd10cf1ec0f270e3beecfe6cc3fba3c adds the new kind of error to
the hardware callback and sends it through the run orchestrator store to
the run orchestrator and thence the engine. This is a convoluted path
because the same callback handles things like the estop, which needs to
be signalled even if there's no engine. We react to asynch errors by:
- Doing nothing, if there's no current run at all (the orchestrator
store swallows the error if there's no run)
- Doing nothing, if the module that got the error does not appear to be
used in the protocol
- This means that all asynchronous errors will be ignored until and
unless somebody runs a load module command, including during protocol
setup. Do we want this?
- Cancelling the run by stopping the engine, if the module that got the
error is used in the protocol

I think that what's here does an adequate job of patching the
_practical_ problem - that if you keep a thermocycler at 4C,
occasionally condensation breaks a thermistor, which probably won't
happen in the preprotocol - but we probably want to enhance the logic to
add more considered reactions in odd situations. I think we have a good
basis for adding that logic, here.

## extending down through the modules
3c252e72bbd10cf1ec0f270e3beecfe6cc3fba3c adds a new kind of callback
down through modules and refactors some stuff along the way. The error
callback (and the disconnected callback) become mandatory, which
simplifies some logic and just removes a complexity doubling; this needs
some test refactors, which is annoying but should improve things.

There was actually already an error callback mechanism in the pollers
themselves; the errors just weren't really hooked up to anything but a
local cache that wouldn't be queried by the engine in the way that we
need. We can link the hardware callback from the previous commit to the
reader callbacks that already exist, and hey presto we can raise
asynchronous errors.

## Testing
~We need to actually have some asynchronous errors happen in modules.
These are pretty annoying to actually create, so we can probably do it
with a hacked-up module firmware to just create an error condition when
you set a temperature in e.g. a tempdeck. Then you can test this by
running a protocol that uses the module, and checking that it fails; and
exercise the preprotocol and out of protocol variants by commanding the
module from the app.~

You can't actually test this all the way through from a module because
none of them express errors in the commands these pollers run. Happily,
this means we can just merge this without doing it (as long as normal
runs work, which they do) because the error won't be triggered.

I plan to do the end to end testing when I implement the modules
actually returning asynch errors in their polling in subsequent work.

## further work
Should probably give the same treatment to module disconnections.
Should probably change what happens when there's an error for a module
during the pre-protocol flow.

Closes EXEC-1705