feat(ci-mkdocs): adding mkdocs to static-deploy method (#19536)
# Overview

Use the new static-deploy scripts to resolve deployment configuration
and with new triggers and logic, run this workflow on `staging-mkdocs*`
and `mkdocs*` tag pushes. This will fully deploy to the respective
environment, staging and prod.

On PR, a push to `edge`, or a push to any `chore_release*` branch:
deploy to sandbox.

> [!NOTE]
>Workflow dispatch handling will be added in a later PR. We have to make
tradeoffs and add significant logic changes to handle overrides and
needed inputs from a workflow dispatch. If there is some issue with a
deployment, we will build and deploy locally to resolve the issues until
we have the workflow dispatch in place.

## Test Plan and Hands on Testing

See that the sandbox deploy works.
To test tag pushes work, we must merge the changes that enable this
workflow and the static-deploy scripts into a `chore_release` branch we
want to push tags at.