refactor(app-shell): extend discovery client to support multiple windows (#19533)
Closes EXEC-1887

Currently, when a secondary window opens, the main window navigates back to /devices. After following the debug trail, the core issue that we registerDiscoveryClient any time we open a secondary window, but the internals of this function clear the discovery cache and start the discovery client anew every time we register.

We don't need (and probably do not ever want) secondary windows to manage the discovery client lifecycle. Let's leave the internals of that management to the main window (the current behavior) and instead, let's just ensure that discovery client correctly dispatches appropriate list updates to secondary windows via a cache mechanism.