feat(shared-data): create `ethanol_80/2.json` and `glycerol_50/2.json` as copies of `1.json` (#19571)
# Overview

This PR introduces the liquid class v2 definition files for `ethanol_80`
and `glycerol_50`. For this PR, they are just direct copies of the v1
files (with `"version": 1` changed to `"version": 2`).

In a future PR, we will add the new liquid class values from
Andy/Mehdi/Thassyo, which will include the values for the new LV96
pipette.

## Test Plan and Hands on Testing

Run CI tests.

## Review requests

@jbleon95 suggested doing this as 2 commits (first copy 1.json ->
2.json, then introduce the new values) to make it easier to see what
changed.

## Risk assessment

Low.