fix(api): improve ramp rate version checking (#19584)
Make the version checking a little more resilient to odd version numbers
(in general, defaulting to ramp-rate-not-allowed if we don't understand
it).

Also, the dev server was running a set temperature with ramp rate; now
that that's not ignored, it's actually not allowed unless the
thermocycler supports it, and we don't have a way for the simulator to
express a module firmware version. So for now just get rid of it. We'll
have to add the ability to specify the firmware version in the simulator
configs.

## testing
- [x] tests don't hang
- [x] `make -C robot-server dev-flex` starts a robot server