(docs): add PCR lid cleanliness standards to Flex manual (#19578)
# Overview

This PR reorganizes the lids section of the thermocycler module info in
the Flex manual. It includes cleanliness and sterility standards for the
single-use lid. It also adds images of the single-use lid and the rubber
seal.

RTC-823

## Test Plan and Hands on Testing

Changes are minor. Make sure links and images work.

## Changelog

Removes a table describing the Thermocycler lids and replaces it with a
dedicated section.

Also, adds 2 images.

## Review requests

n/a

## Risk assessment

Low, docs only.

---------

Co-authored-by: Ed Cormany <edward.cormany@opentrons.com>