refactor(api): Do not hardcode labware versions in deck conflict checker (#19577)
This fixes a hypothetical future bug in the way we check for conflicts
between labware and modules on the deck.

We give special handling to a hard-coded list of labware. We were
specifying each of those labware by its _full URI_, including its
version. That means that the hard-coded list would fall behind whenever
we add new versions of those labware. (Luckily, we have not done that so
far.) With this PR, we just hard-code the namespace and load name parts.