fix(shared-data): Delete incorrect transforms in the Stacker module definition (#19606)
## Overview

This fixes some bad numbers in the Flex Stacker's module definition that
were causing labware to be rendered slightly out of place in the
frontend.

## Details

If the stacker is "in" slot C3, the offset from the front-left corner of
slot C3 to the front-left corner of the shuttle is (161, 0, 31). This is
already correctly recorded in the module definition and deck definition,
and is not changed here.

However, on top of that, the definition had some slot-specific
adjustments. These appear to have been copypasta'd from the
Heater-Shaker's definition. They added a small offset in x and y, and a
large offset in z, moving the stacker quite far under the deck. There
were also OT-2-specific adjustments, which of course don't make sense
because this module is Flex-specific.