fix(protocol-designer): instrumentation to see why getMinXYDimension() is failing (#19626)
# Overview

We have a long-outstanding PD crash when we try to call
`getMinXYDimension()` in `SecondStepsMoveLiquidTools.tsx`, but I can't
figure out why it's failing:
https://opentrons-76.sentry.io/issues/6804827811/?project=4509363266387968

I'm not sure if it's because `formData['aspirate/dispense_labware']` is
blank, or if the labware is not in `labwareEntities`, or if the labware
is in `labwareEntities` but its `def` is missing for some reason.

This PR adds some instrumentation to try to catch the problem earlier,
and give us more information about why it's failing.

## Test Plan and Hands on Testing

Ran PD locally.

## Risk assessment

Low.