fix(protocol-designer): don't crash in volumeTooHigh() if tipRack is null (#19631)
# Overview

I have not been able to reproduce these PD crashes reported in Sentry:
- https://opentrons-76.sentry.io/issues/6867128274/?project=4509363266387968
- https://opentrons-76.sentry.io/issues/6867128191/?project=4509363266387968

But from the stack trace, they seem to be happening because we're
calling `volumeTooHigh()` with a null `tipRack` from the form, which in
turn calls `getPipetteCapacity()` and `getTiprackVolume()`, where we
crash.

Note we previously fixed the crash in `volumeTooHigh()` caused by
`pipette` being null (PR #19353), so we can fix the crash caused by
`tipRack` being null in the same way.

## Test Plan and Hands on Testing

Smoke tested PD running locally.

## Risk assessment

Low.