docs: stacker-hepa-upgrade (#19645)
# Overview

Stacker instruction manual indicates some Stackers need an upgrade to
work with the HEPA module. However, the info is unclear about _who_
performs the upgrade/installs the kit.

Update Stacker manual to clearly indicate installing the upgrade kit is
an Opentrons technician function only. Users are not expected to, and
shouldn't, install the kit.

Flex manual already includes this and does not require a change.

RTC-852

## Test Plan and Hands on Testing

Read it to make sure this makes sense.

## Changelog

Adds a `note` callout to the Stacker compatibility section in
`compliance.md`.

> The upgrade kit must be installed by an Opentrons field service
technician only.

## Review requests

<!--
- What do you need from reviewers to feel confident this PR is ready to
merge?
- Ask questions.
-->

## Risk assessment

Low, docs only.