refactor(protocol-designer): make PD_REQUIRED_APP_VERSION a Vite constant (#19657)
# Overview

Part 1 of 2 for AUTH-2225.

Previously, the Robot Stack version that PD requires was specified in
`ProtocolOverview/ProtocolMetadata.tsx`. This PR makes it a Vite
constant instead (called `_OT_PD_REQUIRED_APP_VERSION_`), so that it's
available at build time.

An upcoming PR will use this constant to retrieve the labware
definitions that are available in that version of the app.

## Test Plan and Hands on Testing

Ran PD locally and looked at the "Required app version" in
http://localhost:5178/#/overview.

## Risk assessment

Low.