feat(protocol-designer): get latest labware versions as of an app version (#19661)
# Overview

Part 2 for AUTH-2225.

This implements a new Vite constant `_OT_PD_LATEST_LABWARE_VERSIONS_`,
which is a map of:
```
{ labware loadname -> highest available version of that labware }
```
as of the app version specified in `REQUIRED_APP_VERSION`.

The map is built by asking git to list all the files in
`shared-data/labware/definitions/2/` as of `REQUIRED_APP_VERSION`. (To
find the files in that app version, we first try looking for a release
tag named `v#.#.#`, and if that doesn't exist, we look for a branch
named `chore_release-#.#.#`.)

You can then refer to `_OT_PD_LATEST_LABWARE_VERSIONS_` from any PD file
to access the map.

@jerader Let me know if the format of the map works for you. I can tweak
the format if you need.

## Test Plan and Hands on Testing

Ran PD locally. I tried `console.log()`ing the new
`_OT_PD_LATEST_LABWARE_VERSIONS_` and it looks reasonable.

## Risk assessment

Low.