chore(hardware-testing): test dynamic pipetting valid on all labware. (#19668)
# Overview
test each well type under the worst case scenario to see if we ever
exceed the z speed while dynamic pipetting.
Using the EM pipette and the 96ch as the worst cases, (the 96 channel
moves more liquid/well/second if there is >65 nozzles in a single well)
and since the worst case for each type of well is the bottom-most
section (well.bottom() + the dead space height) lets just brute force
all of our labware and see if any of them wouldn't work.

Put this test into hardware testing, since it needs the opentrons
package we don't want to add circular dependencies into shared-data so I
couldn't put it there. And it doesn't really test the opentrons package
so I didn't think it made sense there either.
In my mind it is testing our hardware's limits, so hardware-testing it
is.