refactor(app, labware-designer, labware-library, opentrons-ai-client, shared-data): Move all labware images to shared-data (#19088)
# Overview

Relocate labware images from opentrons-ai-client, labware-designer, app,
and labware-library to shared-data

## Test Plan and Hands on Testing

- Tested labware-library, tested the app, tested labware-creator

## Changelog

- Removed duplicate `labware-images.ts` and images folder in the app,
opentrons-ai-client, labware-designer, and labware-library
- Kept an images folder in `labware-designer` that only contains svgs of
labware shapes
- in shared-data, a `labware-images.ts` can be run to generate
`labware-images-generated.ts` . This automatically pairs the labware
load name with the associated image based on the file name. This will
pair labware + the adapter if part of the load name string is found in
the image file name.

## Review requests

- confirm that this is the desired format

## Risk assessment

- need to update some images for better visuals - images are being taken
on 09/17