feat(hardware-testing): dump csv runtime params to the output csv. (#19644)
# Overview
Now that the grav test is a protocol that makes extensive use of runtime
parameters we need a way to track what options were used when running a
particular test.
In the past everything was just committed directly to the code so we
could see the code changes but now we can do most of that work without
making commits.
This will now just dump the csv information into the output csv by
taking the List[List[str]] from protocol ctx and creating sections/lines
and filling them in.