feat(protocol-designer): Module forms show prior module state (#19655)
## Overview

This adds a "Last module state" section to the Thermocycler,
Heater-Shaker, and Temperature Module forms.

Goes towards EXEC-1744, EXEC-1745, and EXEC-1746.

## Changelog

* Add a shared component for the list of "last module state" key-value
blocks. It's modeled as a `<dl>` for semantics.
* Add them to the Thermocycler, Heater-Shaker, and Temperature Module
forms, gated behind a feature flag.
* Wire them up to the output of `step-generation`.

There is one big known issue, EXEC-1899. The displayed state will
accidentally follow whichever step you're hovering. It's supposed to
remain constant, representing just the step that you have open right
now. The same problem existed before for other forms. We're not sure yet
if we're using the wrong hook, or if the right hook is doing the wrong
thing. If we don't have a quick answer in this PR, we can fix it in a
follow-up.

---------

Co-authored-by: Jethary Alcid <66035149+jerader@users.noreply.github.com>