refactor: handle multi-ack commands in serial
The new error status commands in the module have multiple acks in
response to the command itself. In the current implementation, the
serial machinery reads only one command response, which means that the
command after an M411 would get the second ack to the M411, and the
command after that would get the response to the command that got the
M411 response... it's off by one forever.

To fix this we can add a a new serial communication method that handles
these kinds of commands, backed by some extra logic that is also a nice
clean refactor to the serial machinery.