refactor: do not override run errors with estop
This was around to handle the case where the engine was stopped by an
asynchronous estop call; we wouldn't get the estop error and the run
would look successful, so if the engine was stopped asynchronously then
we'd override with an estop error. But now there's multiple asynchronous
error sources, so only do the override if we were going to finish cleanly.