fix(api): set_stored_labware(): preserve prior behavior for lid/adapter namespace/version (#19717)
## Overview

This fixes a backwards compatibility bug in my implementation of #19348
/ EXEC-1750.

## Details

#19348 did two things, in various places:

1. It added new params
   * `adapter_namespace`
   * `adapter_version`
   * `lid_namespace`
   * `lid_version`
2. It adjusted the behavior when those params were omitted

In one of those places, I forgot:

* to raise an error if you tried using the new params on an older
`apiLevel`s
* to preserve the old behavior when the params were omitted

This PR adds that backwards-compatibility behavior, following the
pattern from the rest of #19348.

This is targeting `chore_release-8.7.0`. v8.7.0 is the first robot
software release that would have the changes from #19348.