refactor(api): allow async errors to be set once
We need to reset the run error from a finish() call when we're handling
an async error, even if the ongoing command got cancelled first and set
the error to cancelled; but we also don't want the ongoing command to
finish second and overwrite the async error with the command's cancelled
error. So override the result once. I really hate this.