fix(protocol-designer): Fix newlocation/newLocation typo (#19739)
## Overview

This fixes a logic bug where Protocol Designer was trying to read a
nonexistent `newlocation` property. It should be `newLocation`, with a
capital L.

TypeScript didn't catch this because we're typing `FormData` with `any`.
That's a known problem that we hope we can fix, but it'd be too big for
this PR.

## Test Plan and Hands on Testing

Just CI?

## Risk assessment

Low? It's pretty obvious what this code was trying to do.

Admittedly, I'm not sure what user-facing impact this code has. So I
guess it's possible that this bug was masking some other bug. Is there
anything in particular that we ought to manually test?