feat(api): cancel protocols on module disconnects (#19732)
When you disconnect a module that is used by the protocol, cancel the
protocol.

There's one other weird thing: I've made the details of how run errors
get set much worse, unfortunately.

The way they're intended to be set is that a command fails; something
catches the command failure and sets the run error to be the command's
error; then several things call finish() and that has no errors that can
be carried. But async errors come with finish(), and the order in which
the finish() call comes in relative to the command failure (that is
caused by the cancellation) isn't well defined.

So now let's always set the error from finish, if there is one, and it's
async, overriding any command failure related errors; and let's set it
only once, so that subsequent command failures don't break it.

## testing
- [x] on a flex, run a protocol that uses a module and disconnect the
module; the protocol should cancel
- [x] on a flex, run a protocol that uses a module and disconnect a
different module; the protocol shouldn't care

Closes EXEC-1707
Closes EXEC-861