refactor(protocol-designer): Delete duplicate step form types (#19723)
We had some identical definitions in `.../ProtocolSteps/types.ts` and
`.../ProtocolSteps/StepForm/types.ts`. The duplication looked
accidental. Maybe from a Git merge a while ago, or something?

This deletes `.../ProtocolSteps/types.ts` in favor of
`.../ProtocolSteps/StepForm/types.ts`. The latter appears to have had a
couple more properties defined than the former, so I'm assuming that
means it's newer and better. And it was much more broadly used.