refactor(robot-server, api): genericize FileProvider (#19755)
Closes EXEC-1956

This PR genericizes the FileProvider interface to pave the way for writing more than CSV file types to disk. To that end, FileProvider now takes raw byte data, a MIME type, and command_metadata, which is used to generate the final filename.

FileProviderWrapper is renamed to FileProviderExecutor and now handles the entirety of file naming and writing the file to the correct location. The vision is that shortly, FileProviderExecutor will need to ingest robot-server specific metadata alongside the command_metadata for building the filename, and this step moves us in that direction.