chore: Upgrade Prettier from v2.2.1 to v3.6.2 (#19774)
## Overview

This upgrades Prettier from v2.2.1, which is ~5 years old, to v3.6.2.
Unfortunately, there have been a lot of little tweaks to Prettier's
style in the meantime, so this causes a big diff.

## Risk assessment

Low. Nothing here should incur any behavioral change. I think the risk
is really just that it might cause annoying merge conflicts for a while.