feat(api): add dynamic tracking start and end locations (#19762)
# Overview
This essentially adds the TrackFromLocation and the TrackToLocation to
the protocol engine commands.

It then passes new information down to the OT3Api, replacing the Z
distance with the end point, and updating the motion utilities to create
the target position from that instead of z_distance.

If no TrackToLocation is provided to the engine command then it
preserves the old behavior where it tracks the meniscus.

As is the protocol api just preserves the current behavior, hooking up
the TrackToLocation with that layer is another ticket.

TODO add checking that makes sure the X/Y/Z move is a valid move that
can occur within the timeframe of the plunger move.