feat(robot-server): initialize `images` directory (#19803)
Closes EXEC-1969

For the upcoming camera capture work, we'll need to persist image files on the robot. We don't want to handle saving images in our persistence directory, because doing so opens up issues when performing database migrations.

This commit adds a new robot setting, images_directory, which allows for specifying the directory in which images will be saved. We largely mirror the behavior of persistence_directory here: if no explicit env is passed, we create a temporary directory. By default via settings injection, if the persistence directory is /data/opentrons_robot_server, then the images directory is analogously specified as /data/images.