fix(components): change ListItems colors from green/yellow/red35 to 30 (#19818)
# Overview

Felix and @mmencarelli confirmed that the `ListItem` colors should be
green30/yellow30/red30 instead of green35/yellow35/red35.

(The original green35/yellow35/red35 colors were added back in
PR #14317.)

## Test Plan and Hands on Testing

Looked at the colors.

I couldn't find an easy-to-access place in PD that currently has a
`ListItem` that can be in an error state, but I'm about to add one in an
upcoming PR (for when the user creates a transfer step without selecting
a tiprack).

## Risk assessment

Low.