fix(protocol-designer): require a tip rack in page 0 of the Transfer form (#19821)
# Overview

Continuing to fix the null tiprack issues on Sentry:
- https://opentrons-76.sentry.io/issues/6892806132/?project=4509363266387968
- https://opentrons-76.sentry.io/issues/6914903257/?project=4509363266387968

I believe these PD crashes are happening because we do not check that
you've selected a tiprack in the first page of the Transfer form, and we
let you click `Continue` even if no tiprack is selected.

The code that runs subsequently, such as `getMatchingTipLiquidSpecs()`,
expects to be able to read properties from that tiprack, and fails
because the tiprack is null.

1. The fix is to add the `tiprackRequired` check to
`steplist/formLevel/index.ts`, which didn't have it before.

2. But also, the `tiprackRequired()` implementation was buggy. It was
trying to attach the error to a field called `tiprack`, when we named
the field `tipRack` in the form.

3. I also changed the `TiprackField.tsx` component to highlight the
field in red when it is in an error state, so that users get some
feedback about what is wrong with the form.

After this change, if no tiprack is available and none is selected, we
mark the tiprack field as an error when you click `Continue`, and block
you from going to the next page of the form.

#### Future work

Note that we implemented the component as `<ListItem><StyledText>No
tiprack available</StyledText></ListItem>` when there are no tipracks
available. There's no place to put the error message from
`tiprackRequired()`. That's something we'll have to fix in a future
change.

(Felix also pointed out that the text `No tiprack available` ought to be
red, but we'll have to fix that later too.)

## Test Plan and Hands on Testing

You can test by hand with this attached protocol:
[nulltiprack.json](https://github.com/user-attachments/files/22959657/nulltiprack.json).
The deck in this protocol has no tipracks at all.

Import it into PD, create a Transfer step, and click `Continue`.

Before this change, we would let you continue without a tiprack, and the
`Uncaught TypeError: Cannot read properties of null (reading
'includes')` error will show up in the console. (For some reason, this
error does **not** trigger the whitescreen popup.)

After this change, we will flag the Tiprack field as an error when you
click `Continue`.

## Risk assessment

Low.