test(shared-data): show command when useCommandTypeSummaries test fails (#19833)
# Overview

When `useCommandTypeSummaries.test.tsx` fails, it emits the error
```
AssertionError: expected 'Unknown' not to be 'Unknown'
```
which is not very helpful for explaining what's wrong. (AUTH-2353)

This PR adds the offending command to the error message:
```
captureImage: expected 'Unknown' not to be 'Unknown'
```
which is slightly more informative.

## Test Plan and Hands on Testing

Ran `yarn vitest` locally.

## Risk assessment

Low.