refactor(api, robot-server): Update `data_files` table to support camera enablement (#19835)
Works toward EXEC-1978

The image gallery contains media cards that contain information such as the command associated with the capture image command, the previous command, and the failed command that caused the image capture, if any. The app has no effective way to derive this information consistently and accurately, especially in instances in which the protocol run is non-deterministic.

To that end, the data files table is refactored heavily. The data files table itself now contains relational data concerning the physical, saved-to-disk file. Two new tables are added, an input_data_files and output_data_files table, which provide info when a data file is used as input to a run or when a data file is generated as output to a run.